requirements_full.txt
absl-py==2.1.0
accelerate==0.30.1
alembic==1.13.1
annotated-types==0.7.0
astunparse==1.6.3
blis==0.7.11
Bottleneck @ file:///C:/b/abs_f05kqh7yvj/croot/bottleneck_1707864273291/work
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
catalogue==2.0.10
certifi @ file:///C:/b/abs_35d7n66oz9/croot/certifi_1707229248467/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpathlib==0.16.0
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
colorlog==6.8.2
confection==0.1.4
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
flatbuffers==24.3.25
fsspec @ file:///C:/b/abs_d5jzmndkba/croot/fsspec_1714461594761/work
gast==0.5.4
gmpy2 @ file:///C:/ci/gmpy2_1645455782955/work
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.0
h5py==3.11.0
huggingface_hub @ file:///C:/b/abs_2629p5gdd8/croot/huggingface_hub_1716498314460/work
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
imbalanced-learn==0.12.3
importlib-metadata @ file:///C:/b/abs_c1egths604/croot/importlib_metadata-suite_1704813568388/work
intel-openmp==2021.4.0
Jinja2 @ file:///C:/b/abs_f7x5a8op2h/croot/jinja2_1706733672594/work
joblib @ file:///C:/b/abs_4b02c83r3b/croot/joblib_1713976789617/work
keras==3.3.3
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
Mako==1.3.5
marisa-trie==1.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
mdurl==0.1.2
mkl==2021.4.0
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
ml-dtypes==0.3.2
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
murmurhash==1.0.10
namex==0.0.8
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
nltk==3.8.1
numexpr @ file:///C:/b/abs_5fucrty5dc/croot/numexpr_1696515448831/work
numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp39-cp39-win_amd64.whl#sha256=f284633067fde4bb078a09085a1c9e99a3218c129f5312b7b3d76c9340442a25
opt-einsum==3.3.0
optree==0.11.0
optuna==3.6.1
packaging @ file:///C:/b/abs_cc1h2xfosn/croot/packaging_1710807447479/work
pandas @ file:///C:/b/abs_3awk0iw2ab/croot/pandas_1709590545218/work/dist/pandas-2.2.1-cp39-cp39-win_amd64.whl#sha256=f678af1dc65824be67d5789cb9f76ffef40030e586438eef2c8664c33bbd4c6c
pillow @ file:///C:/b/abs_95ls_qh4c2/croot/pillow_1714398870615/work
preshed==3.0.9
protobuf==4.25.3
psutil==5.9.8
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
pytz @ file:///C:/b/abs_6ap4tsz1ox/croot/pytz_1713974360290/work
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
regex @ file:///C:/b/abs_d5e2e5uqmr/croot/regex_1696515472506/work
requests @ file:///C:/b/abs_474vaa3x9e/croot/requests_1707355619957/work
rich==13.7.1
safetensors @ file:///C:/b/abs_88nwhm1qj3/croot/safetensors_1708633899663/work
scikit-learn @ file:///C:/b/abs_d6_krvz8aa/croot/scikit-learn_1714164798159/work
scipy @ file:///C:/b/abs_2ea4kcjtpg/croot/scipy_1714080455387/work/dist/scipy-1.13.0-cp39-cp39-win_amd64.whl#sha256=900f80f3cad808d3dca1a1d4c888772d33a30ed628e4926cd58e215cfacf3e4d
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
sympy @ file:///C:/b/abs_82njkonm7f/croot/sympy_1701397685028/work
tbb==2021.11.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thinc==8.2.3
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tokenizers @ file:///C:/b/abs_ba1xnwbavr/croot/tokenizers_1708633883417/work
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm @ file:///C:/b/abs_a7h7n45ccq/croot/tqdm_1716395987133/work
transformers @ file:///C:/b/abs_495035xqf4/croot/transformers_1708700635762/work
# Editable Git install with no remote (truthfulness_classifier==0.1)
-e c:\users\alex veskoukis\desktop\satalia\satalia data science challenge\truthfulness_classifier\pythonproject
typer==0.9.4
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///C:/b/abs_8e4z8_gh1l/croot/urllib3_1715636317140/work
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.3
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wrapt==1.16.0
zipp @ file:///C:/b/abs_b0beoc27oa/croot/zipp_1704206963359/work
